[Stack]  - refactor and clean-up
