+++ data-structures/packages/stack/src/index.js
