+++ data-structures/packages/stack/src/index.ts
