+++ features/0.1-Implement-ts-monorepo.md
@@ -1,0 +1,53 @@
+
+## Set up TypeScript
+
+# Implement TS In monorepo
+
+A monorepo is a single repository that is used to manage multiple  projects.
+
+- All we need is already built into npm and TypeScript
+
+
+## Why monorepos ?
+
+- Useful for maintain  multiple interdependent packages in parallel.  
+
+## Structure proposed
+
+```
+my-package-1/
+    ts/
+        src/
+            index.ts
+        util/
+        test/
+            index.js
+    js/
+        src/
+        util/
+        test/
+    /dist-ts
+    /dist-js
+
+my-package-2/
+    ts/
+        src/
+            index.ts
+        util/
+        test/
+    js/
+        src/
+            index.js
+        util/
+        test/
+    /dist-ts
+    /dist-js
+```
+
+## ts-config.js
+
+Ref: 
+
+https://www.typescriptlang.org/docs/handbook/tsconfig-json.html
+
+
